backend/pyproject.toml
[project]
name = "zerokey-backend"
version = "0.1.0"
description = "ZeroKey backend — Django monolith for Malaysian e-invoicing."
requires-python = ">=3.12,<3.13"
dependencies = [
    "django>=5.1,<5.2",
    "djangorestframework>=3.15,<4.0",
    "django-environ>=0.11,<1.0",
    "django-cors-headers>=4.4,<5.0",
    "psycopg[binary]>=3.2,<4.0",
    "redis>=5.0,<6.0",
    "celery[redis]>=5.4,<6.0",
    "django-redis>=5.4,<6.0",
    "gunicorn>=23.0,<24.0",
    "structlog>=24.4,<25.0",
    "python-json-logger>=2.0,<3.0",
    "drf-spectacular>=0.27,<1.0",
    "boto3>=1.35,<2.0",
    "pdfplumber>=0.11,<1.0",
    "anthropic>=0.39,<1.0",
    # EasyOCR for image / scanned-PDF text extraction (Slice 31). Pulls in
    # torch (CPU) + opencv-python-headless + Pillow as transitive deps; the
    # CPU torch wheel is ~200MB and EasyOCR's English language model is
    # downloaded on first use (cached in the easyocr/ directory).
    "easyocr>=1.7,<2.0",
    # RapidOCR (Slice 72) — PaddleOCR's PP-OCR models reorganised to run on
    # ONNX Runtime. Same accuracy as PaddleOCR but ~200MB smaller install
    # (no torch / no paddle), faster cold start, and no GPU deps. Wins at
    # priority 50 over EasyOCR (priority 100) for invoices because of
    # PP-OCR's stronger table-aware detection.
    "rapidocr-onnxruntime>=1.3,<2.0",
    # PDF → image rasterisation for OCR. pypdfium2 is a single Python wheel
    # with no system deps, much leaner than pdf2image+poppler.
    "pypdfium2>=4,<5",
]

[dependency-groups]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-django>=4.9,<5.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.7,<1.0",
    "mypy>=1.13,<2.0",
    "django-stubs[compatible-mypy]>=5.1,<6.0",
    "djangorestframework-stubs>=3.15,<4.0",
    "ipython>=8.28,<9.0",
]

[tool.uv]
# We run inside the Linux Docker container in dev + prod; some deps
# (torch, EasyOCR's transitive Pillow build) don't publish Intel-macOS
# wheels. Declaring required environments tells uv to only resolve for
# the platforms we actually run on, so a no-wheel platform doesn't
# block the lockfile.
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'linux' and platform_machine == 'aarch64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = [
    "E", "W",
    "F",
    "I",
    "B",
    "UP",
    "DJ",
    "S",
    "RUF",
]
ignore = [
    "S101",   # assert (pytest)
    "E501",   # handled by formatter
    "RUF012", # mutable defaults in Meta — idiomatic in Django
    "RUF001", # ambiguous Unicode in code — false-positive on legit non-Latin chars
    "RUF002", # ambiguous Unicode in docstrings — same false-positive surface
    "RUF003", # ambiguous Unicode in comments — same false-positive surface
    "RUF043", # pytest match= metacharacters — works fine for the common case
    "RUF059", # unpacked-but-unused test fixtures — pytest pattern
    "S110",   # try/except/pass — used intentionally for best-effort writes
              # (audit chain best-effort updates, etc.); the alternative is
              # noisier per-call logging without changing behaviour
    "S314",   # ElementTree on untrusted XML — we parse OUR signed XML, not
              # external; a defusedxml swap is a future hardening but not a
              # current vulnerability surface
    "UP042",  # StrEnum — minor 3.11+ optimisation; not worth the churn yet
    "E402",   # module-level imports not at top — Django pattern with
              # conditional imports (csrf_exempt only loaded on the webhook
              # endpoint, etc.) is common and intentional
    "B017",   # blind pytest.raises(Exception) — used for "any exception"
              # signal in a few legacy tests; tightening is follow-up work
    "B904",   # raise without from — readability tradeoff in nested
              # exception flows; can be tightened later
    "F841",   # unused local variable — sometimes intentional in tests for
              # clarity (e.g. master = ... to set up state)
]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["S105", "S106"]
"**/test_*.py" = ["S105", "S106"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
exclude = ["migrations/"]

[tool.django-stubs]
django_settings_module = "zerokey.settings.dev"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "zerokey.settings.test"
python_files = ["test_*.py", "tests.py"]
addopts = "-ra --strict-markers"
testpaths = ["apps", "zerokey"]
